Split logic and animation internal types
