+++ src/game_data_logic_types.h
@@ -1,0 +1,146 @@
+#ifndef SLAYER3D_GAME_DATA_LOGIC_TYPES_H
+#define SLAYER3D_GAME_DATA_LOGIC_TYPES_H
+
+#include <stdbool.h>
+
+#include "script_internal.h"
+#include "slayer3d/game_data.h"
+#include "slayer3d/script.h"
+#include "yyjson.h"
+
+typedef enum game_data_sensor_type
+{
+    GAME_DATA_SENSOR_BOUNDS_EXIT,
+    GAME_DATA_SENSOR_BOUNDS_REFLECT,
+    GAME_DATA_SENSOR_CONTACT_2D,
+    GAME_DATA_SENSOR_HEARING,
+    GAME_DATA_SENSOR_INPUT_PRESSED,
+    GAME_DATA_SENSOR_BRUSH_CONTENTS,
+    GAME_DATA_SENSOR_BRUSH_PERCEPTION,
+    GAME_DATA_SENSOR_PERCEPTION,
+    GAME_DATA_SENSOR_SECTOR,
+    GAME_DATA_SENSOR_VOLUME,
+} game_data_sensor_type;
+
+typedef struct named_signal
+{
+    const char *name;
+    int id;
+} named_signal;
+
+typedef struct named_timer
+{
+    const char *name;
+    float delay;
+    int signal_id;
+    bool repeating;
+    float interval;
+} named_timer;
+
+typedef struct named_action
+{
+    const char *name;
+    int id;
+} named_action;
+
+typedef struct adapter_entry
+{
+    char *name;
+    char *lua_script_id;
+    char *lua_function;
+    slayer3d_script_ref lua_function_ref;
+    slayer3d_game_data_adapter_fn callback;
+    void *userdata;
+} adapter_entry;
+
+typedef struct script_entry
+{
+    const char *id;
+    const char *path;
+    const char *module;
+    const char **dependencies;
+    int dependency_count;
+    slayer3d_script_ref module_ref;
+    bool autoload;
+    bool loading;
+    bool loaded;
+} script_entry;
+
+typedef struct binding_entry
+{
+    struct slayer3d_game_data_runtime *runtime;
+    yyjson_val *actions;
+    int connection_id;
+} binding_entry;
+
+typedef struct sensor_contact_pair_state
+{
+    const char *actor_name;
+    const char *other_actor_name;
+    bool owns_actor_name;
+    bool owns_other_actor_name;
+    bool active;
+    bool seen;
+} sensor_contact_pair_state;
+
+typedef struct sensor_entry
+{
+    game_data_sensor_type type;
+    yyjson_val *json;
+    const char *name;
+    const char *entity;
+    const char *other;
+    const char *entity_tag;
+    const char *other_tag;
+    const char *sector_level;
+    const char *sector;
+    const char *sector_property;
+    int sector_index;
+    const char *action;
+    const char *axis;
+    const char *side;
+    float min_value;
+    float max_value;
+    float threshold;
+    float range;
+    float min_dot;
+    float observer_eye_height;
+    float target_eye_height;
+    const char *yaw_property;
+    slayer3d_vec3 volume_min;
+    slayer3d_vec3 volume_max;
+    int signal_id;
+    yyjson_val *actions;
+    const char *edge;
+    bool was_active;
+    sensor_contact_pair_state *contact_pairs;
+    int contact_pair_count;
+    int contact_pair_capacity;
+} sensor_entry;
+
+typedef struct wave_schedule_entry
+{
+    yyjson_val *schedule;
+    float elapsed;
+    bool initialized;
+} wave_schedule_entry;
+
+typedef struct noise_event_runtime
+{
+    unsigned int id;
+    char key[32];
+    const char *source_actor_name;
+    slayer3d_vec3 position;
+    float radius;
+    float loudness;
+    float remaining_seconds;
+} noise_event_runtime;
+
+typedef struct sensor_actor_list
+{
+    slayer3d_registered_actor **items;
+    int count;
+    int capacity;
+} sensor_actor_list;
+
+#endif
